Make the ServiceSpecificDecorator more generic
